Revert "feat: expiry for prepaid one-off plan items"
