+++ server/src/internal/billing/v2/utils/initFullCustomerProduct/initFullCustomerProduct.ts
@@ -9,7 +9,6 @@ import type { AutumnContext } from "@/honoUtils/HonoEnv";
 import { customerProductHasPaidLicenses } from "@/internal/billing/v2/utils/customerProductHasPaidLicenses.js";
 import { applyExistingStatesToCustomerProduct } from "@/internal/billing/v2/utils/initFullCustomerProduct/applyExisting/applyExistingStatesToCustomerProduct";
 import { generateId } from "@/utils/genUtils";
-import { splitExpiringPurchaseGrants } from "../expiringGrants/splitExpiringPurchaseGrants";
 import { initCustomerEntitlement } from "./initCustomerEntitlement/initCustomerEntitlement";
 import { initCustomerLicenses } from "./initCustomerLicenses/initCustomerLicenses";
 import { initCustomerPrice } from "./initCustomerPrice";
@@ -73,13 +72,6 @@ export const initFullCustomerProductWithBalanceTransitions = ({
 		customerProduct: newFullCustomerProduct,
 		fullProduct,
 		customerLicenseQuantities: initContext.customerLicenseQuantities,
-	});
-
-	// A scheduled product's credits start their clock when it activates, not
-	// when the request was made.
-	splitExpiringPurchaseGrants({
-		customerProduct: newFullCustomerProduct,
-		now: initOptions?.startsAt ?? initContext.now,
 	});
 
 	const balanceTransitionPlan = applyExistingStatesToCustomerProduct({
