+++ server/src/internal/products/product-items/productItemUtils/itemToPriceAndEnt.ts
@@ -173,8 +173,6 @@ export const toFeature = ({
 		usage_limit: null,
 
 		rollover: item.config?.rollover,
-		expiry_duration: item.config?.expiry?.duration,
-		expiry_length: item.config?.expiry?.length,
 		feature_override: item.config?.feature_override,
 	};
 
@@ -243,8 +241,6 @@ const toFeatureAndPrice = ({
 		usage_limit: item.usage_limit || null,
 
 		rollover: item.config?.rollover,
-		expiry_duration: item.config?.expiry?.duration,
-		expiry_length: item.config?.expiry?.length,
 		feature_override: item.config?.feature_override,
 	};
