+++ shared/api/products/items/apiPlanItemV1.ts
@@ -8,7 +8,6 @@ import {
 import { BillingMethod } from "@api/products/components/billingMethod.js";
 import { DisplaySchema } from "@api/products/components/display.js";
 import { ApiPriceProcessorsSchema } from "@api/products/components/processors.js";
-import { EntitlementExpirySchema } from "@models/productModels/durationTypes/entitlementDuration.js";
 import { RolloverExpiryDurationType } from "@models/productModels/durationTypes/rolloverExpiryDurationType.js";
 import { BillingInterval } from "@models/productModels/intervals/billingInterval.js";
 import { ResetInterval } from "@models/productModels/intervals/resetInterval.js";
@@ -189,11 +188,6 @@ export const ApiPlanItemV1Schema = z
 			.meta({
 				internal: true,
 			}),
-
-		expiry: EntitlementExpirySchema.nullish().meta({
-			description:
-				"Purchased units expire this long after each purchase. One-off prepaid consumable items only.",
-		}),
 
 		feature_override: ApiFeatureOverrideSchema.optional().meta({
 			description:
