+++ shared/api/products/items/mappers/planItemV1ToV0.ts
@@ -79,8 +79,6 @@ export function planItemV1ToV0({
 				}
 			: null,
 
-		expiry: item.expiry ?? undefined,
-
 		rollover: item.rollover
 			? {
 					max: item.rollover.max ?? null,
