--- shared/drizzle/0076_solid_maria_hill.sql
@@ -1,1 +1,0 @@
-CREATE INDEX CONCURRENTLY "idx_customer_entitlements_expiring_grants" ON "customer_entitlements" USING btree ("expires_at") WHERE "customer_entitlements"."customer_product_id" IS NOT NULL AND "customer_entitlements"."expires_at" IS NOT NULL;
