+++ shared/drizzle/meta/_journal.json
@@ -533,13 +533,6 @@
       "when": 1788531508069,
       "tag": "0075_gorgeous_hobgoblin",
       "breakpoints": true
-    },
-    {
-      "idx": 76,
-      "version": "7",
-      "when": 1789149879997,
-      "tag": "0076_solid_maria_hill",
-      "breakpoints": true
     }
   ]
 }
