+++ shared/utils/productV2Utils/productItemUtils/mappers/itemToPriceAndEnt.ts
@@ -185,8 +185,6 @@ export const toFeature = ({
 		usage_limit: null,
 
 		rollover: item.config?.rollover,
-		expiry_duration: item.config?.expiry?.duration,
-		expiry_length: item.config?.expiry?.length,
 		feature_override: item.config?.feature_override,
 	};
 
@@ -257,8 +255,6 @@ const toFeatureAndPrice = ({
 		usage_limit: item.usage_limit || null,
 
 		rollover: item.config?.rollover,
-		expiry_duration: item.config?.expiry?.duration,
-		expiry_length: item.config?.expiry?.length,
 		feature_override: item.config?.feature_override,
 	};
