+++ vite/src/views/products/plan/catalog/buildUpdateCatalogPlanParams.ts
@@ -27,21 +27,11 @@ const planItemsToCatalogParams = (
 	items: ApiPlanV1["items"],
 ): NonNullable<UpdateCatalogPlanParamsInput["items"]> =>
 	items.map(
-		({
-			feature,
-			display,
-			reset,
-			price,
-			proration,
-			rollover,
-			expiry,
-			...item
-		}) => ({
+		({ feature, display, reset, price, proration, rollover, ...item }) => ({
 			...omitStripePriceId(item),
 			...(reset ? { reset } : {}),
 			...(price ? { price: omitStripePriceId(price) } : {}),
 			...(proration ? { proration } : {}),
-			...(expiry ? { expiry } : {}),
 			...(rollover
 				? {
 						rollover: {
